관리자 대시보드 API 사용 권한 설정

+++ seboard/src/main/java/com/seproject/admin/role/application/AdminRoleAppService.java
@@ -1,11 +1,16 @@
 package com.seproject.admin.role.application;
 
+import com.seproject.account.account.domain.Account;
 import com.seproject.account.role.domain.Role;
 import com.seproject.account.role.service.RoleService;
-import com.seproject.admin.role.controller.dto.RoleDTO;
+import com.seproject.account.utils.SecurityUtils;
+import com.seproject.admin.dashboard.domain.DashBoardMenu;
+import com.seproject.admin.dashboard.service.AdminDashBoardServiceImpl;
 import com.seproject.admin.role.controller.dto.RoleDTO.CreateRoleRequest;
-import com.seproject.admin.role.controller.dto.RoleDTO.DeleteRoleRequest;
 import com.seproject.admin.role.controller.dto.RoleDTO.RoleResponse;
+import com.seproject.error.errorCode.ErrorCode;
+import com.seproject.error.exception.CustomAccessDeniedException;
+import com.seproject.error.exception.CustomAuthenticationException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,7 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static com.seproject.admin.role.controller.dto.RoleDTO.*;
+import static com.seproject.admin.role.controller.dto.RoleDTO.UpdateRoleRequest;
 
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
@@ -22,7 +27,23 @@ public class AdminRoleAppService {
 
     private final RoleService roleService;
 
+    private final AdminDashBoardServiceImpl dashBoardService;
+
+    //TODO : 추후 AOP로 변경 필요
+    private void checkAuthorization(){
+        Account account = SecurityUtils.getAccount()
+                .orElseThrow(() -> new CustomAuthenticationException(ErrorCode.NOT_LOGIN, null));
+
+        DashBoardMenu dashboardmenu = dashBoardService.findDashBoardMenuByUrl(DashBoardMenu.ROLE_MANAGE_URL);
+
+        if(!dashboardmenu.authorize(account.getRoles())){
+            throw new CustomAccessDeniedException(ErrorCode.ACCESS_DENIED, null);
+        }
+    }
+
     public List<RoleResponse> findAll(){
+        checkAuthorization();
+
         List<Role> roles = roleService.findAll();
 
         return roles.stream()
@@ -32,16 +53,22 @@ public class AdminRoleAppService {
 
     @Transactional
     public Long createRole(CreateRoleRequest request) {
+        checkAuthorization();
+
         return roleService.createRole(request.getName(), request.getDescription(), request.getAlias());
     }
 
     @Transactional
     public void deleteRole(Long roleId) {
+        checkAuthorization();
+
         roleService.deleteRole(roleId);
     }
 
     @Transactional
     public void updateRole(Long roleId, UpdateRoleRequest request) {
+        checkAuthorization();
+
         //TODO
         roleService.updateRole(roleId,request.getName(),request.getDescription(),request.getAlias());
     }
